requirements.txt
django==4.2.11
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
pandas==2.2.2
gunicorn==21.2.0
django-environ==0.11.2
